rename shared test fixture module to satisfy the push-gate's tests.rs pattern
test_support.rs's push was blocked by the repo's pre-push guideline hook
(unwrap()/expect() in production code): the hook exempts files by path
(tests\.rs$ or /tests/), and this new file didn't match despite being
#[cfg(test)]-gated and containing only test-only content. Renamed to
shared_fixture_tests.rs, which does.
