+++ src/resolver/shared_fixture_tests.rs
